Started refactoring allocation and buffer pool code
The current code uses several allocators and buffer pools. This commit
establishes one common allocator for physical memory blocks, and one
buffer pool that can be used with it. The result is better code reuse,
less code, and buffer pool & allocator behavior that is more conformant
to the GStreamer design.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>