ipu: Open IPU FD once, globally, for all blitter instances, with refcount
It is preferable to open/close the IPU FD once globally, as opposed to
once for each blitter instance, since this circumvents some bugs in the
IPU driver with multiple open FDs.

Also, the refcount is introduced, to make sure the FD is closed only after
all blitter instances are shut down and all allocators are finalized.
The latter is important to prevent IPU_FREE ioctl's from being called
after the FD was closed.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>