common: add base classes for 2D blitters and blitter-based elements
This is useful for IPU, G2D, PXP based video sinks and transform elements.
The base classes take care of most GStreamer details, and the derived
ones can focus on the specific blitter details.

Available base classes are:

* GstImxBaseBlitter
* GstImxBlitterVideoSink
* GstImxBlitterVideoTransform

GstImxBaseBlitter instances are used inside the video sink and transform
elements. These use the blitter for blitting the video frames on screen
or to a new output frame.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>