eglvivsink: Add support for window-x/y coords in Wayland
Wayland does not provide a way for controlling the absolute placement of
Windows in the compositor. It does, however, provide a "subcompositor"
interface which allows absolute positioning of subsurfaces within a parent
surface.

The patch moves the video surface to a "subsurface" which is positioned
within a main, transparent surface. Care is taken to properly configure
the input region of the window so that events are only captured within
the proper subsurface region.

The "window_x_coord" and "window_y_coord" parameters are only honored at
the sink configuration time so it is not (yet) possible to reposition
the video while live.

[ Patch sent by Edison Fernandez <edison.fernandez@ridgerun.com> ]

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>