imxeglvivsink: add an android platform.
Creates a new platform: 'android' introducing egl_platform_android.c and the
configure option '--egl-platform=android'. The platform, file inspired from the
framebuffer platform, uses EGL and the Android native window support to provide a
display.

Contrary to the framebuffer support, the EGL window surface is created from the
injected native window and not from scratch. It is mandatory as the only viable
way do display something on Android is to get a window from the SurfaceFlinger.
It won't be consistent to draw directly in the framebuffer.

Plus, G2D based plugins are disabled as the 2D coree is used by Android for the
display framework composition and cannot be re-used for anything else.

Based on the work of Éric Chaloin.
Reviewed by: Gaëtan Morin <gaetan.morin@pulse-origin.com>