imxeglvivsink: remove GLESv2 VIVANTE link deps.
Removes the link dependency to libGLESv2_VIVANTE as it is not compatible with
the Android NDK. The NDK only provides the standard EGL/GLES{,v2}, the usage of
any vendor extension require the caller to retrieves symbols from the library
using the standard EGL primitive 'eglGetProcAddress()'. This allow binaries to
be compatible with all Android platforms at link level.

Reviewed by: Gaëtan Morin <gaetan.morin@pulse-origin.com>