eglvivsink: Make use of wayland subsurface configurable
This commit introduced the usage of a subsurface:

commit 854a11493b7a4887350acc38916aa6890b9fbdb8
Author: Carlos Rafael Giani <dv@pseudoterminal.org>
Date:   Mon Jul 6 08:01:05 2015 +0200

    eglvivsink: Add support for window-x/y coords in Wayland

The side effect is that the sink consumes about twice as much
memory, which might be quite a lot on a high res screen.

So by adding a property to disable the subsurface the memory
can be saved, but x/y will not work as earlier.

Signed-off-by: Richard Röjfors <richard@puffinpack.se>