audio: Rempve objdump config scan
The objdump-based scan during configuration is problematic and fragile.
So, instead, use a hardcoded table to load them at runtime. (Since caps
have to be associated to the plugins, it is necessary to hardcode their
names to assign the caps properly.)

Also, if the attempt to dlopen() the plugin with full path fails, add
another attempt with the filename only, which works in some cases.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>