vpu/encoder_base: add zerocopy GRAY8 to h264/h263/mpeg4 encoding
This patch adds zerocopy GRAY8 to h264/h263/mpeg4 conversion by using
a constant memory zone populated with the value '128' for the Cb and Cr
planes that are requested by the h264/h263 and mpeg4 encoders, but
not provided by GRAY8.

As it is needed by the h264, h263 and mpeg4 encoders, and to avoid
code duplication, this has been added to encode_base.c and is
activated by each encoder when needed.

Because of the API of libimxvpuapi, I had to set cb_offset and cr_offset
to the difference between the physical address of the new zone and
the physical address of the Y plane, which is a hack I admit, but
this works perfectly.

Signed-off-by: Philippe De Muyter <phdm@macqel.be>