vpu: Add memory tag to VPU framebuffer meta
The VPU framebuffer meta is primarily a memory meta, since its whole
purpose is to make sure VPU framebuffers are returned to the VPU pool.
This implies that it is always accompanied by a imxphysmeta. In some cases
however it was possible that GstMeta transformations got rid of the
physmeta but passed through the VPU meta, causing a segfault. Example:

.. imxvpudec ! videoconvert ! "video/x-raw, format=BGRx" ! imxg2dvideosink

videoconvert would drop the physmeta but not the VPU meta, causing the
pipeline to crash.

Add a memory tag to make sure BOTH physmeta and VPU meta are dropped.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>