Major rewrite for a version 2.0.0, introducing i.MX8 support
This is a rewrite, based on the new libimxdmabuffer library, the new
libimxvpuapi library version 2.0.0, and GStreamer 1.14. The i.MX8 Hantro
VPU is supported. i.MX6 SoCs still supported as well.

GStreamer 1.14 is now a minimum requirement to make integration with
glimagesink and other elements easier and to be able to remove legacy
cruft and keep the code cleaner.

The code was reorganized into gst-libs/ and ext/ subdirectories to match
the GStreamer directory structure.

Also, meson is now the build system in use to ease integration with
GStreamer build setups such as Cerbero.

Note that this is not yet a fully finished 2.0.0 release. It is still WIP,
though the VPU en/decoder elements should be fully functional.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>