Update decoder.c
I have found that frequent format changes causes leaks of the vpu_decoder->decoder structure, and there are only 32.  The reason is that if we come into this function with a NULL context, we skip the code that calls decoder close.  It happens that the context can be NULL but the vpu_decoder->decoder can still be open.