imx2d: Add overlay composition meta support to GstImx2dVideoTransform
This adds code to render overlay meta compositions into output frames.
with the imx2d blitter. This improves performance and provides better
output since the overlays are drawn with the output frame's resolution.

It also checks if downstream can handle such a meta; if so, the overlays
aren't rendered into the output, and the meta is passed through instead.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>