v4l2videosrc: Rework code to use DMA-BUF if available
This replaces the USERPTR based approach that was used with non-mxc_v4l2
devices. When used with newer imx kernels, the mmap() syscall to get a
virtual memory pointer for USERPTR hangs. It is currently not known why.
To work around this, if the allocator can produce DMA-BUF memory, use
DMA-BUF memory mode. The non-DMA-BUF case is currently unsupported, and
will be addressed in a later commit. mxc_v4l2 based devices are not
affected by this; these still use the USERPTR method with the special
imx hack for passing physical addresses to the driver.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>