imx-gst-libs: common: Replace dup() with memory ref'ing in dmabuf uploader
Using dup() was causing an FD leak because the dup'ed FD was not being
closed correctly. But also, just using dup() risks data races because
producers can still write to the same dma-buf memory that consumers are
processing. Instead, ref the input dmabuf gstmemory to let the rest of
the pipeline know that the gstmemory is in use.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>