imx-gst-libs: video: Adjust video meta before doing a CPU based frame copy
This fixes cases where the very first frame might be distorted but the
followup ones are okay. gst_video_frame_map() uses the videometa values
from the source and destination buffers if present. The buffer pool
produces buffers with a videometa, but it is one that is not containing
aligned stride values. Consequently, the videometa must be adjusted
_before_ the gst_video_frame_map() call to prevent that one from using
unaligned stride values.

In practice, followup frames may be okay, since the buffers that get
returned to the pool contain videometas with previously adjusted stride
values by then. But by adjusting first, it is ensured that they are
always already adjusted prior to the frame mapping and copy.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>